requirements.txt
# AI Trading Signal System - Requirements

# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0

# SmartAPI
SmartApi>=1.3.0
pyotp>=2.8.0

# Technical Analysis
# Option 1: TA-Lib (recommended, but requires C library)
# Install TA-Lib C library first: https://ta-lib.org/install/
# Then: pip install TA-Lib

# Option 2: ta library (pure Python, easier to install)
ta>=0.10.2

# Machine Learning
xgboost>=1.7.0

# Telegram
requests>=2.28.0

# Logging
logzero>=1.7.0

# Scheduling
schedule>=1.2.0

# Optional: For better performance
# numba>=0.56.0
